Make config('auth.guards.web.expire') always return an int
The config value can return an int or a string, depending on
whether the envvar is set or the default is used.

Move the cast to the config definition, so it's always an int
and we don't get RedundantCast issues from psalm.

+++ config/auth.php
@@ -60,7 +60,7 @@ return [
         'web' => [
             'driver'    => 'session',
             'provider'  => 'users',
-            'expire'    => env( 'AUTH_TOKEN_EXPIRE', 43200 ), // optional token expiration time, in minutes (30 days is the default)
+            'expire'    => (int) env( 'AUTH_TOKEN_EXPIRE', 43200 ), // optional token expiration time, in minutes (30 days is the default)
         ],
 
         'api' => [
